DB refactor
