+++ packages/core/src/messages/acknowledgement/index.ts
@@ -5,7 +5,7 @@ import { Challenge } from '../packet/challenge'
 import { Hash, Signature, PublicKey } from '@hoprnet/hopr-core-ethereum'
 import PeerId from 'peer-id'
 import { serializeToU8a, u8aSplit, u8aToHex } from '@hoprnet/hopr-utils'
-import { UnAcknowledgedTickets } from '../../dbKeys'
+import { UnAcknowledgedTickets } from '../../db'
 
 /**
  * This class encapsulates the message that is sent back to the relayer
